requirements.txt
rasa[spacy]
requests
python-dotenv